LCD-44026 allow CorsOrigin annotation in applications
