+++ modules/util/client-extension-util-spring-boot/build.gradle
@@ -4,6 +4,7 @@ targetCompatibility = "1.8"
 dependencies {
 	api group: "org.springframework.boot", name: "spring-boot-starter-oauth2-client", version: "2.7.18"
 	api group: "org.springframework.boot", name: "spring-boot-starter-oauth2-resource-server", version: "2.7.18"
+	api group: "org.springframework.boot", name: "spring-boot-starter-web", version: "2.7.18"
 	api group: "org.springframework.boot", name: "spring-boot-starter-webflux", version: "2.7.18"
 
 	implementation group: "com.liferay", name: "com.liferay.petra.string", version: "latest.release"
@@ -18,4 +19,5 @@ dependencies {
 	testImplementation group: "org.mock-server", name: "mockserver-client-java", version: "3.10.8"
 	testImplementation group: "org.mock-server", name: "mockserver-netty", version: "3.10.8"
 	testImplementation group: "org.mockito", name: "mockito-core", version: "1.10.8"
+	testImplementation group: "org.springframework.boot", name: "spring-boot-test-autoconfigure", version: "2.7.18"
 }
